// types/pointsTableDashBoardType.tsx
type PointsTableDashBoardType = {
  teamId: number;
  position: number;
  teamName: string;
  teamLogo: string;
  matchesPlayed: number;
  matchesWon: number;
  matchesLost: number;
  matchesDrawn: number;
  netRunRate: number;
  forTeams: string;
  againstTeams: string;
  recentForm: string;
  points: number;
};

export type { PointsTableDashBoardType };